Build the good suffix table in linear time
Previously, this was O(m^2) where m is the length of the
pattern. We can build it in linear time by continuing to
slide the pattern over itself when a suffix match is found
rather than restarting for each index.

This commit also makes it so that an exact match on the
entire string can still make use of the good suffix rule
to skip characters. For example, an exact match on the
string ABBBA lets us skip 4 characters:

   ABBBA
       ABBBA
       ^
       |
next possible match